+++ src/features/toggle_secret_door.js
@@ -1,4 +1,5 @@
 import {settingsKey} from "../settings.js"
+import {updateSynchronizedDoors} from "./synchronized_doors.js";
 
 // Toggles between normal and secret doors
 export function onDoorLeftClick(event) {
@@ -6,7 +7,13 @@ export function onDoorLeftClick(event) {
 	if (game.settings.get(settingsKey, "toggleSecretDoors") && event.data?.originalEvent?.ctrlKey && game.user.isGM) {
 		const types = CONST.WALL_DOOR_TYPES
 		const newtype = this.wall.data.door === types.DOOR ? types.SECRET : types.DOOR
-		this.wall.update({door: newtype})
+		const updateData = {door: newtype}
+		const synchronizationGroup = this.wall.data.flags.smartdoors?.synchronizationGroup
+		if (game.settings.get(settingsKey, "synchronizedDoors") && synchronizationGroup && this.wall.data.flags.smartdoors?.synchronizeSecretStatus)
+			updateSynchronizedDoors(updateData, synchronizationGroup)
+		else
+			this.wall.update(updateData)
+
 		return true
 	}
 	return false