+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## In development
+### New features
+- Synchronized doors can now be configured to synchronize their secret door status as well
+
 ## v1.2.4
 ### Bugfix
 - Fixed a race condition that may cause doors to not be properly synchronized across scenes